requirements.txt
git+https://github.com/django-nonrel/django-nonrel@master
argparse==1.2.1
distribute==0.6.24
django-extensions==0.8
django-mongodb-engine==0.4.0
django-tastypie==0.9.11
djangotoolbox==0.9.2
gunicorn==0.14.2
mimeparse==0.1.3
pymongo==2.1.1
pystache==0.5.2
python-dateutil==1.5
wsgiref==0.1.2
